% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes.R
\name{fit_hawkes}
\alias{fit_hawkes}
\title{Using nelder mead to estimate Hawkes process MLEs}
\usage{
fit_hawkes(
  params_init,
  realiz,
  windowT,
  windowS,
  maxit,
  exponential_kernel = TRUE,
  kernel = NULL,
  optimized = TRUE,
  trace = 0,
  poisson_flag = FALSE,
  zero_background_region = NULL,
  ...
)
}
\arguments{
\item{params_init}{list with elements,}

\item{realiz}{list with elements, n,lon,lat,t}

\item{windowT}{vector with elements, start,end}

\item{windowS}{os.win}

\item{maxit}{maxit for optim}

\item{trace}{trace for optim}

\item{poisson_flag}{flag for if this process is really a pure poisson process}

\item{...}{other arguments to pass to optim}
}
\value{
optim fit object
}
\description{
Fit a hawkes process by maximizing th log likelihood with the nelder mead algorithm
}
