% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes.R
\name{loglik_hawk}
\alias{loglik_hawk}
\title{Log likelihood for spatio temporal log likelihood}
\usage{
loglik_hawk(
  params,
  realiz,
  windowT,
  windowS,
  optimized = F,
  exponential_kernel = TRUE,
  kernel = NULL,
  dists = NULL,
  zero_background_region = NULL,
  density_approx = TRUE,
  numeric_integral = FALSE
)
}
\arguments{
\item{params}{list with elements mu,alpha,beta,K}

\item{realiz}{list with elements, n,lon,lat,t}

\item{windowT}{vector with elements, start,end}

\item{windowS}{spatstat.geom owin object or an object that can be coerced as such}

\item{dists}{list with elements space_dist,time_dist}

\item{zero_background_region}{spatstat.geom owin object}

\item{density_approx}{boolean to determine if we are using the density approximation}

\item{numeric_integral}{boolean to determine if we are using the numeric integral}
}
\value{
numerical loglikelihood value
}
\description{
Calculates the log likelihood of a hawkes process realization
}
