% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes.R
\name{sim_hawkes}
\alias{sim_hawkes}
\title{Simulate Spatio Temporal Hawkes Processes}
\usage{
sim_hawkes(
  params,
  windowT,
  windowS,
  background_realization = NULL,
  filtration = NULL,
  optimized = TRUE
)
}
\arguments{
\item{params}{list with elements,}

\item{windowT}{vector with elements, start,end}

\item{windowS}{owin object}

\item{background_realization}{list with elements, n,lon,lat,t which is used as the background if supplied}

\item{filtration}{\itemize{
\item points outside the time window that can trigger new points still (ie. a past history)
}}

\item{optimized}{boolean to determine if we are using the optimized version (i.e. slight faster due to vectorization)}
}
\value{
list with t,x,y for point process
}
\description{
simulating a recursive Hawkes process with lambda(t,x,y) = mu + SUM g(t_i,x_i,y_i) for g some kernel function
where:
\itemize{
\item g(t) is kenel function with parameters alpha and beta
\item mu is the background rate,
\item K is the average number of aftershocks generated by an event.
}
}
